+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 *[CalVer, YY.month.patch](https://calver.org/)*
 
+## Future
+- Added TRIO105 check for not immediately `await`ing async trio functions.
+
 ## 22.7.3
 - Added TRIO102 check for unsafe checkpoints inside `finally:` blocks